aitools-api/requirements.txt
Flask==2.3.2
pytesseract==0.3.10
Pillow==10.0.1
Werkzeug==2.3.7
Flask-Cors==3.0.10
requests